xen/arm: ffa: reclaim shared memory on guest destroy
When an FF-A enabled guest is destroyed it may leave behind memory
shared with SPs. This memory must be reclaimed before it's reused or an
SP may make changes to memory used by a new unrelated guest. So when the
domain is teared down add FF-A requests to reclaim all remaining shared
memory.

SPs in the secure world are notified using VM_DESTROYED that a guest has
been destroyed. An SP is supposed to relinquish all shared memory to allow
reclaiming the memory. The relinquish operation may need to be delayed if
the shared memory is for instance part of a DMA operation.

The domain reference counter is increased when the first FF-A shared
memory is registered and the counter is decreased again when the last
shared memory is reclaimed. If FF-A shared memory registrations remain
at the end of of ffa_domain_teardown() a timer is set to try to reclaim
the shared memory every second until the memory is reclaimed.

A few minor style fixes with a removed empty line here and an added new
line there.

Signed-off-by: Jens Wiklander <jens.wiklander@linaro.org>
Reviewed-by: Bertrand Marquis <bertrand.marquis@arm.com>
---
v5:
- Move the put_domain() call after the spinlock section in dec_ctx_shm_count()
- Move the get_knownalive_domain() call into the else statement in
  inc_ctx_shm_count() to make sure it's only performed if ctx->shm_count
  is to be increased

v4:
- Retry FFA_MSG_SEND_VM_DESTROYED on the returned errors
  FFA_RET_INTERRUPTED and FFA_RET_RETRY even after all shared memory
  handles has been reclaimed.
- Retry ffa_mem_reclaim() only on temporary returned errors, permanent
  errors like FFA_RET_INVALID_PARAMETERS is dealt with in recovery mode
  instead since the SPMC isn't expected to use that error under normal
  circumstances.

v3:
- Mentioning in the commit message that there are some style fixes
- Addressing review comments
- Refactor the ffa_domain_teardown() path to let
  ffa_domain_teardown_continue() do most of the work.

v2:
- Update commit message to match the new implementation
- Using a per domain bitfield to keep track of which SPs has been notified
  with VM_DESTROYED
- Holding a domain reference counter to keep the domain as a zombie domain
  while there still is shared memory registrations remaining to be reclaimed
- Using a timer to retry reclaiming remaining shared memory registrations